Handle special case for array access.
Array access in code is to be specified as `a[i]`, but is forbidden as
`get a i`, because get is an axiomatized-only function, thus cannot be
used in code. However, before we patch Why theories with our (probably
desugared) ones, we need to support Why3's sugared notation for array
access.

This handles such a notation for array access. Array modification is not
handled, because Why3's notation destructively modifies the array, and
our IR does not allow that. Thus, we will need a non-destructive "set"
operation for that.