Add a --why3 flag to invoke why3 after processing
This flag should improve the user friendlyness by automatically invoking
the Why3 IDE after finishing processing each file. This only works on
the batch file processing workflow (and not on the stdout or multifile
parsing mechanisms).

This is a development aid, but later a console-based why3 flag with
strategy support should be a more useful use case scenario.

GitHub: closes #3