Remove debugging :printv of theories
A debugging printv was introduced in commit
b065ac4ee9bb26260970c9905c4fe6783b5d99fa for getting the theories that
were being read by the program. This is not very useful information, was
presented in a not very user-friendly format, and is now deleted from
the output of the program.

git-svn-id: https://dalila.sip.ucm.es/svn_caviart/representacion-intermedia/vc-gen/clir2mlw@1528 7ef80a5c-f9c9-4c3c-9a70-3e98a7bd2970